feat: add basic timeblock sync

+++ apps/web/src/app/api/meet-together/plans/[planId]/timeblocks/[timeblockId]/route.ts
@@ -7,7 +7,6 @@ export const dynamic = 'force-dynamic';
 interface Params {
   params: {
     timeblockId: string;
-    userType: string;
   };
 }
 
@@ -37,15 +36,44 @@ export async function PUT(
 }
 
 export async function DELETE(
-  _: Request,
-  { params: { timeblockId: id, userType } }: Params
+  req: Request,
+  { params: { timeblockId: id } }: Params
 ) {
   const supabase = createRouteHandlerClient({ cookies });
+
+  const data = await req.json();
+
+  const guestId = data.guestId;
+  const passwordHash = data.passwordHash;
+
+  if (guestId && !passwordHash)
+    return NextResponse.json(
+      { message: 'Password hash required for guest timeblocks' },
+      { status: 400 }
+    );
+
+  if (passwordHash && !guestId)
+    return NextResponse.json(
+      { message: 'Guest id required for password hash' },
+      { status: 400 }
+    );
+
+  const userType = guestId ? 'guest' : 'user';
+
+  const {
+    data: { user },
+  } = await supabase.auth.getUser();
+
+  if (userType === 'user' && !user)
+    return NextResponse.json({ message: 'Unauthorized' }, { status: 401 });
+
+  const userId = userType === 'user' ? user?.id : guestId;
 
   const { error } = await supabase
     .from(`meet_together_${userType}_timeblocks`)
     .delete()
-    .eq('id', id);
+    .eq('id', id)
+    .eq('user_id', userId);
 
   if (error) {
     console.log(error);
